requirements.txt
# Web framework
flask>=3.0
flask-socketio>=5.3
flask-sqlalchemy>=3.1
gunicorn>=22.0
eventlet>=0.35

# Blockchain
web3>=6.0
eth-account>=0.10

# HTTP & async
requests>=2.31
aiohttp>=3.9

# Scheduling & background tasks
apscheduler>=3.10

# Database
sqlalchemy>=2.0

# Environment & config
python-dotenv>=1.0

# Data processing
pandas>=2.0
numpy>=1.24

# Utilities
retry>=0.9
cachetools>=5.0
